The gate scripts share one vocabulary, and the duplicate check can finally see them
`test_there_is_one_copy_of_each_helper` excluded `test_`, `run_` and `validate_` under one
sentence: test modules repeat fixtures on purpose. That reason covers test modules. It does not
cover a gate script, which is not a fixture and imports like any other module -- and 75 of them
carrying 855 definitions sat outside every duplicate check in the tree.

The exclusion is now split rather than dropped, because the two halves of the file want different
corpora. A gate script's module-scope constants ARE its own configuration: two readiness gates
each naming REQUIRED_DOC_SNIPPETS for their own document is not drift, and widening the constant
and pattern checks reported seven such names as if it were. A copied FUNCTION body is a copied
helper wherever it sits. So `BODY_CHECK_PREFIXES` excludes only `test_`, and the body check alone
reads the gates: 217 modules to 292.

It found five copied bodies, every one a `require_*` shape check written out twice --
require_string_set, require_int_map, require_int_at_least and require_distribution_count in both
context gates, require_snippets in both readiness gates. They live once now, in
`tools/matrixark_validation_requirements`, together with the rest of the vocabulary those four
scripts had each spelled out: load_report, require_bool, require_int_equal, require_int_between
and require_map_keys_at_least.

Eighteen definitions become nine. `fail`, `validate_report` and `main` stay where they are -- each
says something different per gate.

Two of the moved bodies were NOT identical and the differences are recorded rather than smoothed
over: `require_bool` returned True in one gate and None in the other, raising the same message
from the same condition, and every call is in statement position; `load_report` opened a handle in
one and read the text in the other. The shared copy takes the returning and the handle-opening
forms.

`require_snippets` reads a module-scope ROOT. The shared module computes it the same way, from
`tools/`, so `parents[1]` is the same directory it was -- which is the check this file's own
history says to make before moving a body that reads a constant.

Verified behaviourally, not just structurally: both context gates run against a ladder of six
synthetic reports that walks the vocabulary -- a non-object root, an empty object, a wrong type, a
failed cross-field sum, a short string set, a bad int map -- and both readiness gates run against
the repository. Twelve runs, byte-identical stdout, stderr and exit code before and after.
