+++ tools/validate_metaserver_production_meta_management.py
@@ -12,6 +12,15 @@ from __future__ import annotations
 
 from pathlib import Path
 
+try:
+    from tools.matrixark_validation_requirements import (
+        require_snippets,
+    )
+except ModuleNotFoundError:  # Direct script execution from tools/.
+    from matrixark_validation_requirements import (
+        require_snippets,
+    )
+
 
 ROOT = Path(__file__).resolve().parents[1]
 
@@ -81,18 +90,6 @@ REQUIRED_DOC_SNIPPETS = (
     "Server heartbeat/liveness feeds scheduler repair decisions",
     "No-majority writes fail closed",
 )
-
-
-def require_snippets(path: Path, snippets: tuple[str, ...], label: str) -> int:
-    if not path.exists():
-        raise SystemExit(f"{label}: missing {path.relative_to(ROOT)}")
-    text = path.read_text(encoding="utf-8", errors="ignore")
-    missing = [snippet for snippet in snippets if snippet not in text]
-    if missing:
-        raise SystemExit(
-            f"{label}: {path.relative_to(ROOT)} missing snippets: {', '.join(missing)}"
-        )
-    return len(snippets)
 
 
 def main() -> None:
